backend/app/requirements.txt
langchain 
langchain-community 
langchainhub 
bs4
langchain-google-vertexai
pypdf
python-dotenv
langchain_google_genai
faiss-cpu